Refactor viewport into separate file

+++ src/camera.rs
@@ -138,9 +138,9 @@ impl CameraController {
         }
     }
 
-    pub fn handle_mouse(&mut self, mouse_dx: f64, mouse_dy: f64) {
-        self.rotate_horizontal = mouse_dx as f32;
-        self.rotate_vertical = mouse_dy as f32;
+    pub fn handle_mouse(&mut self, mouse_dx: f32, mouse_dy: f32) {
+        self.rotate_horizontal = mouse_dx;
+        self.rotate_vertical = mouse_dy;
     }
 
     pub fn handle_scroll_pixels(&mut self, scroll_pixels: f32) {
